[#97] :sparkles: refactor: FileInput 모듈화로 인해 컴포넌트 임포트 방식으로 적용하여 사용하도록 변경
