+++ client/src/components/Shared/Modal/ReviewModal.tsx
@@ -1,11 +1,11 @@
-import React, { useRef, useState } from 'react';
+import React from 'react';
 import Rating from '@/components/Shared/Rating';
 import Title from '@/components/Shared/Title';
 import ModalLayout from './ModalLayout';
 import * as S from './styles';
-import { UploadSVG } from '@/assets/svgs';
 import Button from '@/components/Shared/Button';
-import Textarea from '../Input/Textarea';
+import { FileInput, Textarea } from '@/components/Shared/Input';
+import useFileInput from '@/hooks/useFileInput';
 
 interface ReviewModalProps {
   toggleModal: () => void;
@@ -14,30 +14,13 @@ interface ReviewModalProps {
 // TODO: Input 타입을 조금 더 만들어야 하겠군뇨 호호호
 
 const ReviewModal = ({ toggleModal }: ReviewModalProps) => {
-  const [fileImg, setFileImg] = useState<string[]>([]);
-  const [isError, setIsError] = useState(false);
-  const fileInput = useRef<HTMLInputElement>(null);
-
-  const handleFileUploader = () => {
-    fileInput.current?.click();
-  };
-
-  const handleFileUpload = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const {
-      target: { files },
-    } = e;
-    for (let i = 0; i < (files?.length ?? 0); i++) {
-      const file = files?.[i];
-      const preview = URL.createObjectURL(file);
-
-      if (fileImg.length === 3) {
-        setIsError(true);
-        return;
-      }
-
-      setFileImg([...fileImg, preview]);
-    }
-  };
+  const {
+    fileRef,
+    fileImgs,
+    isError,
+    handleClickOnFileInput,
+    handleUploadFile,
+  } = useFileInput();
 
   return (
     <ModalLayout toggleModal={toggleModal}>
@@ -52,31 +35,15 @@ const ReviewModal = ({ toggleModal }: ReviewModalProps) => {
         <S.Form>
           <Title level={5}>사진 업로드 (최대 3장)</Title>
 
-          <div className="input-file-button" onClick={handleFileUploader}>
-            <UploadSVG width={100} height={100} />
-            <span className="helper-text">
-              클릭 또는 드래그&드랍으로 파일을 추가하세요!
-            </span>
-            <input
-              ref={fileInput}
-              id="file-upload"
-              type="file"
-              onChange={handleFileUpload}
-              multiple
-              hidden
-            />
-          </div>
-          {fileImg.length !== 0 && (
-            <div className="preview-wrapper">
-              {fileImg.map((file) => (
-                <img key={file} src={file} alt="미리보기 이미지" />
-              ))}
-            </div>
-          )}
-
-          {isError && (
-            <p className="error-text">3장 이상의 사진은 업로드 할 수 없어요!</p>
-          )}
+          <FileInput
+            fileRef={fileRef}
+            name="image-uploader"
+            fileImgs={fileImgs}
+            isError={isError}
+            handleClickOnFileInput={handleClickOnFileInput}
+            handleUploadFile={handleUploadFile}
+            hidden
+          />
         </S.Form>
 
         <S.Form>
